Update Snake. Add comments. Update version.
Update Snake with full functionality for updating its own position,
adding a tail for when the eatApple method is called. Added ArrayLists
to store the snake position and an int Array to store the previous tail
position.

Add comments with dates to all of the classes for clarity.

Update version to 0.0.3.