Update Collision, Engine, Snake, and version.
Update Collision with comments to describe each parameter for its
methods.

Update Engine with variables for the Engine to track timing. Update the
game Engine loop which loops at the ups variable's value. Added methods
to update the snake locations, check for collision events with all game
objects, and stop the game Engine thread when called.

Update snake with constants for the direction it is moving and the snake
head index.

Update version.