Add Renderer and EngineEvents. Update Engine, Main, Window, and version.
Add Renderer to draw game objects to the JPanel.

Add EngineEvents to pass events and game updates to the Window.

Update Engine with flags for the game state.

Update Main to initialize Engine first in order to set the initial game
state prior to rendering.

Update Window to draw Renderer.

Update version to 0.2.0.