Update Engine, Input, EngineEvents, and InputEvents. Update version.
Update Engine to account for the game state that it is currently in and
update the game accordingly.

Update Input to check for the space bar being pressed when the game
state is in RESET or GAME_OVER.

Update InputEvents to pass the play flag for the Engine.

Update EngineEvents to pass the location data for the Apple and the
Snake, as well as the score.

Update version to 0.2.1.